+++ AssetsFlow.Client.React/src/Components/Journal/TradeComposite/Controls/TradeNotifications.jsx
@@ -3,7 +3,7 @@ import Notifications from "@components/Notifications";
 import * as Constants from "@constants/journalConstants";
 
 function TradeNotifications({ tradeComposite }) {
-  const hasMissingContent = tradeComposite[Constants.TRADE_ISPENDING];
+  const hasMissingContent = tradeComposite[Constants.HasMissingContent];
   
   if (!hasMissingContent) {
     return null;