+++ AssetsFlow.Client.React/src/Components/Notifications.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Badge, ThemeIcon, Tooltip } from "@mantine/core";
+import { Badge, Tooltip } from "@mantine/core";
 import { TbProgressAlert } from "react-icons/tb";
 
 function Notifications({ shortText, expandedText }) {
@@ -11,7 +11,7 @@ function Notifications({ shortText, expandedText }) {
         variant="light"
         radius="md"
         leftSection={<TbProgressAlert size={14} />}
-        className="hover:bg-red-400 hover:text-white transition-colors"
+        className="mx-1 hover:bg-red-400 hover:text-white transition-colors"
       >
         {shortText}
       </Badge>